refactor : Product 모델 재사용 역할을 명확하기 위해 경로 수정 및 네이밍 변경
