--- app/src/main/java/nextstep/shoppingcart/model/product/Product.kt
@@ -1,16 +1,0 @@
-package nextstep.shoppingcart.model.product
-
-import java.text.DecimalFormat
-
-data class Product(
-    val name: String,
-    val imageUrl: String,
-    val price: String,
-    val productId: String
-) {
-    val formattedPrice: String
-        get() = numberMoneyFormat.format(price.toInt()) + "원"
-
-}
-
-val numberMoneyFormat = DecimalFormat("#,###")
